Start a different browse layout:
- get rid of the enormous list of scheme cards, and move the buttons
  into the intervention map popup
- group all of the schemes stuff and treat as a layer
- move the layers to the left
